다중 파일 업로드 구현

+++ insider/src/main/java/com/kh/insider/repo/MemberRepo.java
@@ -22,5 +22,5 @@ public interface MemberRepo {
 	public void changePassword(MemberDto memberDto);
 	
 	// 닉네임
-	String nick(int memberNo);
+	String nick(long memberNo);
 }
